Implementing SCOE comms for BalBot

+++ src/Arduino - AS6/BalBot/utils.h
@@ -10,6 +10,6 @@
 
 int deadband(int value, int deadband);
 int limit(int value, int min, int max);
-double removeDeadband(double value, double deadband, double range);
-double limit(double value, double min, double max);
+double removeDeadband(float value, float deadband, float range);
+double limit(float value, float min, float max);
 #endif /* UTILS_H_ */
